requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
audioread==2.1.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gast==0.3.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.1
matplotlib==3.2.1
mutagen==1.44.0
numba==0.48.0
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.2.0
pandas==1.0.3
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.23.1
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
sklearn-crfsuite==0.3.6
SoundFile==0.10.3.post1
tabulate==0.8.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-probability==0.10.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.44.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1